requirements.txt
# torch==2.7.0+cu12.8

qwen-vl-utils>=0.0.14
datasets
requests
Pillow
numpy
pymilvus
vllm
python-dotenv
openai
transformers
click
tqdm
huggingface_hub
rank-bm25>=0.2.2
jieba>=0.42.1